+++ src/app/app.routes.ts
@@ -1,7 +1,7 @@
 import { Routes } from "@angular/router"
 
 import { WelcomePage } from "./pages/welcome/welcome"
-import { ProsodyPage } from "./pages/new-poem/step-1/prosody"
+import { ProsodyPage } from "./pages/new-poem/step-1/ProsodyPage"
 import { MivesForms } from "./pages/new-poem/step-2/mives"
 import { WeightsForms } from "./pages/new-poem/step-3/weights"
 import { PoemPage } from "./pages/poem/poem.component"
